Dev -> Main 로그인 기능

+++ pages/api/post/delete.js
@@ -1,17 +1,29 @@
 import { connectDB } from "@/util/database";
 import { ObjectId } from "mongodb";
+import { getServerSession } from "next-auth";
+import { authOptions } from "../auth/[...nextauth]";
 
 export default async function handler(req, res) {
+  let session = await getServerSession(req, res, authOptions);
+  console.log("session : ", session);
   try {
     // 메소드 DELETE로 하면 BODY가 안보내짐
     if (req.method === "POST") {
       console.log("delete body : ", req.body);
       const client = await connectDB;
       const db = client.db("dbname");
-      let result = await db
+
+      let findUser = await db
         .collection("clname")
-        .deleteOne({ _id: new ObjectId(req.body) });
-      res.status(200).json("삭제성공");
+        .findOne({ _id: new ObjectId(req.body) });
+
+      console.log("findUser : ", findUser);
+      if (session.user.email === findUser.author) {
+        let result = await db
+          .collection("clname")
+          .deleteOne({ _id: new ObjectId(req.body) });
+        res.status(200).json("삭제성공");
+      } else console.log("다른 계정입니다");
     }
   } catch (error) {
     console.error(error);
